handle nulls outside numba
